requirements.txt
python-dotenv  
pandas
pyarrow